Refactoring and one bugfix
