+++ loss.py
@@ -1,0 +1,120 @@
+#!/usr/bin/env python3
+
+'''
+Module for all the loss related classes.
+'''
+
+import geopandas as gpd
+import pandas as pd
+
+
+class LossCellList():
+    '''
+    List with the spatial cells with the loss data.
+    '''
+    def __init__(self, loss_cells):
+        self._loss_cells = loss_cells
+
+    def get_loss_cells(self):
+        '''
+        Returns a list of spatial cells with the loss data.
+        '''
+        return self._loss_cells
+
+    def append(self, loss_cell):
+        '''
+        Append a loss cell.
+        There is no logic to merge the cells.
+        '''
+        self._loss_cells.append(loss_cell)
+
+    def to_dataframe(self):
+        '''
+        Converts the loss cell list to a dataframe.
+        '''
+        series = [x.to_series() for x in self._loss_cells]
+        dataframe = pd.DataFrame(series)
+        return gpd.GeoDataFrame(dataframe, geometry=dataframe['geometry'])
+
+
+class LossCell():
+    '''
+    Spatial cell with loss data.
+    '''
+
+    def __init__(self, gid, name, geometry, loss_value, loss_unit):
+        self._gid = gid
+        self._name = name
+        self._geometry = geometry
+        self._loss_value = loss_value
+        self._loss_unit = loss_unit
+
+    def get_gid(self):
+        '''
+        Returns the gid.
+        '''
+        return self._gid
+
+    def get_name(self):
+        '''
+        Returns the name of the cell.
+        '''
+        return self._name
+
+    def get_geometry(self):
+        '''
+        Returns the geometry of the cell.
+        '''
+        return self._geometry
+
+    def get_loss_value(self):
+        '''
+        Returns the computed loss value for this cell.
+        '''
+        return self._loss_value
+
+    def get_loss_unit(self):
+        '''
+        Returns the unit for the loss in this cell.
+        '''
+        return self._loss_unit
+
+    def to_series(self):
+        '''
+        Converts this cell to a pandas series.
+        '''
+        series = pd.Series({
+            'gid': self._gid,
+            'name': self._name,
+            'geometry': self._geometry,
+            'loss_value': self._loss_value,
+            'loss_unit': self._loss_unit,
+        })
+        return series
+
+    @classmethod
+    def from_transition_cell(cls, transition_cell, loss_provider):
+        '''
+        Creates the loss cell from a transition cell and the
+        loss provider.
+        '''
+        loss_value = 0
+
+        for transition in transition_cell.get_transitions():
+            single_loss_value = loss_provider.get_loss(
+                schema=transition_cell.get_schema(),
+                taxonomy=transition.get_taxonomy(),
+                from_damage_state=transition.get_from_damage_state(),
+                to_damage_state=transition.get_to_damage_state()
+            )
+            n_loss_value = single_loss_value * transition.get_n_buildings()
+
+            loss_value += n_loss_value
+
+        return cls(
+            gid=transition_cell.get_gid(),
+            name=transition_cell.get_name(),
+            geometry=transition_cell.get_geometry(),
+            loss_value=loss_value,
+            loss_unit=loss_provider.get_unit()
+        )
