+++ test_exposure.py
@@ -1,18 +1,27 @@
 #!/usr/bin/env python3
 
-import os
-import json
+'''
+Test classes for the exposure.
+'''
+
 import unittest
 
-import geopandas as gpd
 import pandas as pd
 from shapely import wkt
 
-import extendedexposure
+import exposure
 
-class TestNewExposureFormat(unittest.TestCase):
+
+class TestExposure(unittest.TestCase):
+    '''
+    Testclass for the exposur.
+    '''
 
     def test_exposure_cell_list_from_simple_dataframe(self):
+        '''
+        Tests the conversion from an exposure cell list
+        to a dataframe (simple - "old" - format).
+        '''
         geometry1 = wkt.loads('POINT(14 51)')
         geometry2 = wkt.loads('POINT(15 52)')
 
@@ -24,7 +33,7 @@ class TestNewExposureFormat(unittest.TestCase):
             'W+WS/H:1,2_D1': [50.0, 55.0]
         })
         schema = 'SARA_v1.0'
-        exposure_cell_list = extendedexposure.ExposureCellList.from_simple_dataframe(
+        exposure_cell_list = exposure.ExposureCellList.from_simple_dataframe(
             schema=schema,
             dataframe=dataframe
         )
@@ -33,16 +42,28 @@ class TestNewExposureFormat(unittest.TestCase):
 
         recreated_dataframe = exposure_cell_list.to_simple_dataframe()
 
-        self.assertTrue(all(dataframe['geometry'] == recreated_dataframe['geometry']))
-        self.assertTrue(all(dataframe['name'] == recreated_dataframe['name']))
-        self.assertTrue(all(dataframe['gc_id'] == recreated_dataframe['gc_id']))
+        self.assertTrue(all(
+            dataframe['geometry'] == recreated_dataframe['geometry']
+        ))
+        self.assertTrue(all(
+            dataframe['name'] == recreated_dataframe['name']
+        ))
+        self.assertTrue(all(
+            dataframe['gc_id'] == recreated_dataframe['gc_id']
+        ))
 
-        self.assertTrue(all(dataframe['W+WS/H:1,2'] == recreated_dataframe['W+WS/H:1,2_D0']))
-        self.assertTrue(all(dataframe['W+WS/H:1,2_D1'] == recreated_dataframe['W+WS/H:1,2_D1']))
-
-
+        self.assertTrue(all(
+            dataframe['W+WS/H:1,2'] == recreated_dataframe['W+WS/H:1,2_D0']
+        ))
+        self.assertTrue(all(
+            dataframe['W+WS/H:1,2_D1'] == recreated_dataframe['W+WS/H:1,2_D1']
+        ))
 
     def test_exposure_cell_list_from_dataframe(self):
+        '''
+        Tests the conversion from an exposure cell list
+        to a dataframe (more complex).
+        '''
         geometry1 = wkt.loads('POINT(14 51)')
         geometry2 = wkt.loads('POINT(15 52)')
 
@@ -52,7 +73,7 @@ class TestNewExposureFormat(unittest.TestCase):
             'geometry': [geometry1, geometry2],
             'expo': [
                 {
-                    'id': [ 'AREA # 13301', 'AREA # 13301'],
+                    'id': ['AREA # 13301', 'AREA # 13301'],
                     'Region': ['Colina', 'Colina'],
                     'Taxonomy': ['W+WS/H:1,2', 'W+WS/H:1,2'],
                     'Dwellings': [107.5, 107.5],
@@ -63,7 +84,7 @@ class TestNewExposureFormat(unittest.TestCase):
                     'Damage': ['D0', 'D1']
                 },
                 {
-                    'id': [ 'AREA # 13302', 'AREA # 13302'],
+                    'id': ['AREA # 13302', 'AREA # 13302'],
                     'Region': ['Quilpue', 'Quilpue'],
                     'Taxonomy': ['W+WS/H:1,2', 'W+WS/H:1,2'],
                     'Dwellings': [107.5, 107.5],
@@ -76,7 +97,7 @@ class TestNewExposureFormat(unittest.TestCase):
             ]
         })
         schema = 'SARA_v1.0'
-        exposure_cell_list = extendedexposure.ExposureCellList.from_dataframe(
+        exposure_cell_list = exposure.ExposureCellList.from_dataframe(
             schema=schema,
             dataframe=dataframe
         )
@@ -85,15 +106,23 @@ class TestNewExposureFormat(unittest.TestCase):
 
         recreated_dataframe = exposure_cell_list.to_dataframe()
 
-        self.assertTrue(all(dataframe['geometry'] == recreated_dataframe['geometry']))
-        self.assertTrue(all(dataframe['name'] == recreated_dataframe['name']))
-        self.assertTrue(all(dataframe['gid'] == recreated_dataframe['gid']))
-        self.assertTrue(all(dataframe['expo'] == recreated_dataframe['expo']))
-
-
-
+        self.assertTrue(all(
+            dataframe['geometry'] == recreated_dataframe['geometry']
+        ))
+        self.assertTrue(all(
+            dataframe['name'] == recreated_dataframe['name']
+        ))
+        self.assertTrue(all(
+            dataframe['gid'] == recreated_dataframe['gid']
+        ))
+        self.assertTrue(all(
+            dataframe['expo'] == recreated_dataframe['expo']
+        ))
 
     def test_exposure_cell_list(self):
+        '''
+        Tests the exposure cell list.
+        '''
         geometry1 = wkt.loads('POINT(14 51)')
         series1 = pd.Series({
             'name': 'Colina',
@@ -103,7 +132,10 @@ class TestNewExposureFormat(unittest.TestCase):
             'W+WS/H:1.2_D1': 50.0,
         })
         schema = 'SARA_v1.0'
-        exposure_cell1 = extendedexposure.ExposureCell.from_simple_series(schema, series1)
+        exposure_cell1 = exposure.ExposureCell.from_simple_series(
+            schema=schema,
+            series=series1
+        )
 
         geometry2 = wkt.loads('POINT(15 52)')
         series2 = pd.Series({
@@ -114,14 +146,23 @@ class TestNewExposureFormat(unittest.TestCase):
             'W+WS/H:1.2_D1': 30.0,
         })
         schema = 'SARA_v1.0'
-        exposure_cell2 = extendedexposure.ExposureCell.from_simple_series(schema, series2)
+        exposure_cell2 = exposure.ExposureCell.from_simple_series(
+            schema=schema,
+            series=series2
+        )
 
-        exposure_cell_list = extendedexposure.ExposureCellList(exposure_cells=[exposure_cell1, exposure_cell2])
+        exposure_cell_list = exposure.ExposureCellList(
+            exposure_cells=[exposure_cell1, exposure_cell2]
+        )
 
         self.assertEqual(2, len(exposure_cell_list.get_exposure_cells()))
 
-
     def test_exposure_cell_from_simple_series(self):
+        '''
+        Tests the conversion from an exposure cell list
+        to a series (with columns for the taxonomy + damage states
+        and the values of this columns for the number of buildings).
+        '''
         geometry = wkt.loads('POINT(14 51)')
         series = pd.Series({
             'name': 'Colina',
@@ -131,9 +172,12 @@ class TestNewExposureFormat(unittest.TestCase):
             'W+WS/H:1.2_D1': 50.0,
         })
         schema = 'SARA_v1.0'
-        exposure_cell = extendedexposure.ExposureCell.from_simple_series(schema, series)
+        exposure_cell = exposure.ExposureCell.from_simple_series(
+            schema=schema,
+            series=series
+        )
 
-        self.assertEqual('SARA_v1.0',  exposure_cell.get_schema())
+        self.assertEqual('SARA_v1.0', exposure_cell.get_schema())
         self.assertEqual('CHL.14.1.1_1', exposure_cell.get_gid())
         self.assertEqual('Colina', exposure_cell.get_name())
         self.assertEqual(geometry, exposure_cell.get_geometry())
@@ -145,16 +189,19 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual(0, first_tax.get_damage_state())
         second_tax = exposure_cell.get_taxonomies()[1]
         self.assertEqual(1, second_tax.get_damage_state())
-        
 
     def test_exposure_cell_from_series(self):
+        '''
+        Tests the conversion from an exposure cell list
+        to a series.
+        '''
         geometry = wkt.loads('POINT(14 51)')
         series = pd.Series({
             'name': 'Colina',
             'gid': 'CHL.14.1.1_1',
             'geometry': geometry,
             'expo': {
-                'id': [ 'AREA # 13301', 'AREA # 13301'],
+                'id': ['AREA # 13301', 'AREA # 13301'],
                 'Region': ['Colina', 'Colina'],
                 'Taxonomy': ['W+WS/H:1,2', 'W+WS/H:1,2'],
                 'Dwellings': [107.5, 107.5],
@@ -167,9 +214,9 @@ class TestNewExposureFormat(unittest.TestCase):
         })
 
         schema = 'SARA_v1.0'
-        exposure_cell = extendedexposure.ExposureCell.from_series(schema, series)
+        exposure_cell = exposure.ExposureCell.from_series(schema, series)
 
-        self.assertEqual('SARA_v1.0',  exposure_cell.get_schema())
+        self.assertEqual('SARA_v1.0', exposure_cell.get_schema())
         self.assertEqual('CHL.14.1.1_1', exposure_cell.get_gid())
         self.assertEqual('Colina', exposure_cell.get_name())
         self.assertEqual(geometry, exposure_cell.get_geometry())
@@ -183,7 +230,10 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual(1, second_tax.get_damage_state())
 
     def test_exposure_cell(self):
-        tdb1 = extendedexposure.TaxonomyDataBag(
+        '''
+        Test of an exposure cell.
+        '''
+        tdb1 = exposure.TaxonomyDataBag(
             schema='SARA_v1.0',
             taxonomy='W+WS/H:1,2',
             damage_state=0,
@@ -195,7 +245,7 @@ class TestNewExposureFormat(unittest.TestCase):
             population=446.5,
             name='Colina'
         )
-        tdb2 = extendedexposure.TaxonomyDataBag(
+        tdb2 = exposure.TaxonomyDataBag(
             schema='SARA_v1.0',
             taxonomy='W+WS/H:1,2',
             damage_state=1,
@@ -208,15 +258,15 @@ class TestNewExposureFormat(unittest.TestCase):
             name='Colina'
         )
         geometry = wkt.loads('POINT(14 51)')
-        exposure_cell = extendedexposure.ExposureCell(
+        exposure_cell = exposure.ExposureCell(
             schema='SARA_v1.0',
             gid='CHL.14.1.1_1',
             name='Colina',
             geometry=geometry,
-            taxonomies =[tdb1, tdb2]
+            taxonomies=[tdb1, tdb2]
         )
-        
-        self.assertEqual('SARA_v1.0',  exposure_cell.get_schema())
+
+        self.assertEqual('SARA_v1.0', exposure_cell.get_schema())
         self.assertEqual('CHL.14.1.1_1', exposure_cell.get_gid())
         self.assertEqual('Colina', exposure_cell.get_name())
         self.assertEqual(geometry, exposure_cell.get_geometry())
@@ -225,16 +275,46 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual(51, lat)
         self.assertEqual(2, len(exposure_cell.get_taxonomies()))
 
+    def test_extract_damage_state_from_taxonomy_damage_state_string(self):
+        '''
+        Test to extract the damage state from a string with
+        taxonomy and damage state.
+        '''
+        self.assertEqual(
+            0, exposure.extract_damage_state_from_taxonomy_damage_state_string(
+                'W+WS/H:1,2'
+            )
+        )
+        self.assertEqual(
+            2,
+            exposure.extract_damage_state_from_taxonomy_damage_state_string(
+                'W+WS/H:1,2_D2'
+            )
+        )
 
     def test_extract_taxonomy_from_taxonomy_damage_state_string(self):
-        self.assertEqual(0, extendedexposure.extract_damage_state_from_taxonomy_damage_state_string('W+WS/H:1,2'))
-        self.assertEqual(2, extendedexposure.extract_damage_state_from_taxonomy_damage_state_string('W+WS/H:1,2_D2'))
-
-    def test_extract_taxonomy_from_taxonomy_damage_state_string(self):
-        self.assertEqual('W+WS/H:1,2', extendedexposure.extract_taxonomy_from_taxonomy_damage_state_string('W+WS/H:1,2'))
-        self.assertEqual('W+WS/H:1,2', extendedexposure.extract_taxonomy_from_taxonomy_damage_state_string('W+WS/H:1,2_D2'))
+        '''
+        Test to extract the taxonomy from a string with
+        taxonomy and damage state.
+        '''
+        self.assertEqual(
+            'W+WS/H:1,2',
+            exposure.extract_taxonomy_from_taxonomy_damage_state_string(
+                'W+WS/H:1,2'
+            )
+        )
+        self.assertEqual(
+            'W+WS/H:1,2',
+            exposure.extract_taxonomy_from_taxonomy_damage_state_string(
+                'W+WS/H:1,2_D2'
+            )
+        )
 
     def test_taxonomy_data_bag_from_simple_series(self):
+        '''
+        Test to read the taxonomy data from a simple
+        series.
+        '''
         series = pd.Series({
             'W+WS/H:1,2_D3': 100.0,
             'name': 'Colina',
@@ -244,7 +324,11 @@ class TestNewExposureFormat(unittest.TestCase):
         })
 
         schema = 'SARA_v1.0'
-        tdb = extendedexposure.TaxonomyDataBag.from_simple_series(series, 'W+WS/H:1,2_D3', schema)
+        tdb = exposure.TaxonomyDataBag.from_simple_series(
+            series=series,
+            key='W+WS/H:1,2_D3',
+            schema=schema
+        )
 
         self.assertEqual('SARA_v1.0', tdb.get_schema())
         self.assertEqual('W+WS/H:1,2', tdb.get_taxonomy())
@@ -252,14 +336,19 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual(100.0, tdb.get_n_buildings())
         self.assertEqual('Colina', tdb.get_name())
 
-
-
     def test_remove_prefix_d_for_damage_state(self):
-        self.assertEqual(0, extendedexposure.remove_prefix_d_for_damage_state('D0'))
-        self.assertEqual(1, extendedexposure.remove_prefix_d_for_damage_state('D1'))
-        self.assertEqual(10, extendedexposure.remove_prefix_d_for_damage_state('D10'))
+        '''
+        Test to remove the prefix d of the damage state strings.
+        '''
+        self.assertEqual(0, exposure.remove_prefix_d_for_damage_state('D0'))
+        self.assertEqual(1, exposure.remove_prefix_d_for_damage_state('D1'))
+        self.assertEqual(10, exposure.remove_prefix_d_for_damage_state('D10'))
 
     def test_taxonomy_data_bag_from_series(self):
+        '''
+        Test to create a taxonomy data bag from a
+        series.
+        '''
         series = pd.Series({
             'id': 'AREA # 13301',
             'Region': 'Colina',
@@ -274,8 +363,9 @@ class TestNewExposureFormat(unittest.TestCase):
 
         schema = 'SARA_v1.0'
 
-        tdb = extendedexposure.TaxonomyDataBag.from_series(
-                series, schema)
+        tdb = exposure.TaxonomyDataBag.from_series(
+            series, schema
+        )
 
         self.assertEqual('SARA_v1.0', tdb.get_schema())
         self.assertEqual('W+WS/H:1,2', tdb.get_taxonomy())
@@ -291,7 +381,10 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual('Colina', tdb.get_name())
 
     def test_taxonomy_data_bag(self):
-        tdb = extendedexposure.TaxonomyDataBag(
+        '''
+        Test of the taxonomy data bag.
+        '''
+        tdb = exposure.TaxonomyDataBag(
             schema='SARA_v1.0',
             taxonomy='W+WS/H:1,2',
             damage_state=0,
@@ -318,7 +411,14 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual('Colina', tdb.get_name())
 
     def test_fill_cell_step_by_step(self):
-        tdb = extendedexposure.TaxonomyDataBag(
+        '''
+        Test to have an empty exposure cell
+        and to add taxonomy data bags step by step
+        (as it will be done in the update
+        prodecure by applying the fragility functions
+        and the intensities).
+        '''
+        tdb = exposure.TaxonomyDataBag(
             schema='SARA_v1.0',
             taxonomy='W+WS/H:1,2',
             damage_state=0,
@@ -332,7 +432,7 @@ class TestNewExposureFormat(unittest.TestCase):
         )
 
         geometry = wkt.loads('POINT(14 51)')
-        exposure_cell = extendedexposure.ExposureCell(
+        exposure_cell = exposure.ExposureCell(
             schema='SARA_v1.0',
             name='Colina',
             gid='CHL.14.1.1_1',
@@ -352,7 +452,7 @@ class TestNewExposureFormat(unittest.TestCase):
         first_tax = exposure_cell.get_taxonomies()[0]
         self.assertEqual(200.0, first_tax.get_n_buildings())
 
-        tdb2 = extendedexposure.TaxonomyDataBag(
+        tdb2 = exposure.TaxonomyDataBag(
             schema='SARA_v1.0',
             taxonomy='W+WS/H:1,2',
             damage_state=1,
@@ -372,6 +472,12 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual(100.0, second_tax.get_n_buildings())
 
     def test_to_empty_exposure_cell(self):
+        '''
+        Tests the creation of an empty
+        exposure cell from an existing one.
+        Should still have the same
+        name, id and geometry, but no taxonomy data.
+        '''
         geometry1 = wkt.loads('POINT(14 51)')
         series1 = pd.Series({
             'name': 'Colina',
@@ -381,7 +487,10 @@ class TestNewExposureFormat(unittest.TestCase):
             'W+WS/H:1.2_D1': 50.0,
         })
         schema = 'SARA_v1.0'
-        exposure_cell1 = extendedexposure.ExposureCell.from_simple_series(schema, series1)
+        exposure_cell1 = exposure.ExposureCell.from_simple_series(
+            schema=schema,
+            series=series1
+        )
 
         empty_exposure_cell = exposure_cell1.without_taxonomies()
 
@@ -389,133 +498,6 @@ class TestNewExposureFormat(unittest.TestCase):
         self.assertEqual('CHL.14.1.1_1', empty_exposure_cell.get_gid())
         self.assertEqual(geometry1, empty_exposure_cell.get_geometry())
         self.assertEqual([], empty_exposure_cell.get_taxonomies())
-
-    def test_exposure_to_transition_cell(self):
-        geometry1 = wkt.loads('POINT(14 51)')
-        series1 = pd.Series({
-            'name': 'Colina',
-            'gc_id': 'CHL.14.1.1_1',
-            'geometry': geometry1,
-            'W+WS/H:1,2': 100.0,
-            'W+WS/H:1.2_D1': 50.0,
-        })
-        schema = 'SARA_v1.0'
-        exposure_cell1 = extendedexposure.ExposureCell.from_simple_series(schema, series1)
-        transition_cell1 = extendedexposure.TransitionCell.from_exposure_cell(exposure_cell1)
-
-        self.assertEqual(0, len(transition_cell1.get_transitions()))
-
-        transition_to_add = extendedexposure.Transition(
-            schema='SARA_v1.0',
-            taxonomy='W+WS/H:1,2',
-            from_damage_state=0,
-            to_damage_state=1,
-            n_buildings=10,
-        )
-        transition_cell1.add_transition(transition_to_add)
-
-        self.assertEqual(1, len(transition_cell1.get_transitions()))
-        first_transition = transition_cell1.get_transitions()[0]
-
-        self.assertEqual(schema, first_transition.get_schema())
-        self.assertEqual('W+WS/H:1,2', first_transition.get_taxonomy())
-        self.assertEqual(0, first_transition.get_from_damage_state())
-        self.assertEqual(1, first_transition.get_to_damage_state())
-        self.assertEqual(10, first_transition.get_n_buildings())
-
-        transition_cell1.add_transition(transition_to_add)
-
-        self.assertEqual(1, len(transition_cell1.get_transitions()))
-        first_transition = transition_cell1.get_transitions()[0]
-        self.assertEqual(20, first_transition.get_n_buildings())
-
-        self.assertEqual('Colina', transition_cell1.get_name())
-        self.assertEqual('CHL.14.1.1_1', transition_cell1.get_gid())
-        self.assertEqual(geometry1, transition_cell1.get_geometry())
-
-        transition_to_add2 = extendedexposure.Transition(
-            schema='SARA_v1.0',
-            taxonomy='W+WS/H:1,2',
-            from_damage_state=0,
-            to_damage_state=2,
-            n_buildings=10,
-        )
-        transition_cell1.add_transition(transition_to_add2)
-
-        transition_list = extendedexposure.TransitionCellList([transition_cell1])
-        transition_dataframe = transition_list.to_dataframe()
-
-        self.assertEqual('Colina', transition_dataframe['name'][0])
-        self.assertEqual('CHL.14.1.1_1', transition_dataframe['gid'][0])
-        self.assertEqual(geometry1, transition_dataframe['geometry'][0])
-        self.assertEqual(schema, transition_dataframe['schema'][0])
-        
-        transition_element = pd.DataFrame(transition_dataframe['transitions'][0])
-
-        self.assertTrue('taxonomy' in transition_element.columns)
-        self.assertTrue('from_damage_state' in transition_element.columns)
-        self.assertTrue('to_damage_state' in transition_element.columns)
-        self.assertTrue('n_buildings' in transition_element.columns)
-
-        self.assertEqual(2, len(transition_element['taxonomy']))
-        self.assertTrue(all(transition_element['taxonomy'] == 'W+WS/H:1,2'))
-        self.assertTrue(all(transition_element['from_damage_state'] == 0))
-        self.assertTrue(any(transition_element['to_damage_state'] == 1))
-        self.assertTrue(any(transition_element['to_damage_state'] == 2))
-
-
-    def test_exposure_to_loss_cell(self):
-        geometry1 = wkt.loads('POINT(14 51)')
-        series1 = pd.Series({
-            'name': 'Colina',
-            'gc_id': 'CHL.14.1.1_1',
-            'geometry': geometry1,
-            'W+WS/H:1,2': 100.0,
-            'W+WS/H:1.2_D1': 50.0,
-        })
-        schema = 'SARA_v1.0'
-        exposure_cell1 = extendedexposure.ExposureCell.from_simple_series(schema, series1)
-        transition_cell1 = extendedexposure.TransitionCell.from_exposure_cell(exposure_cell1)
-
-        transition_to_add = extendedexposure.Transition(
-            schema='SARA_v1.0',
-            taxonomy='W+WS/H:1,2',
-            from_damage_state=0,
-            to_damage_state=1,
-            n_buildings=10,
-        )
-
-        transition_cell1.add_transition(transition_to_add)
-        loss_cell1 = extendedexposure.LossCell.from_transition_cell(
-            transition_cell1,
-            DummyLossProvider()
-        )
-
-        self.assertEqual('$', loss_cell1.get_loss_unit())
-        self.assertEqual(10, loss_cell1.get_loss_value())
-        self.assertEqual('Colina', loss_cell1.get_name())
-        self.assertEqual('CHL.14.1.1_1', loss_cell1.get_gid())
-        self.assertEqual(geometry1, loss_cell1.get_geometry())
-
-        loss_list = extendedexposure.LossCellList([loss_cell1])
-
-        loss_dataframe = loss_list.to_dataframe()
-
-        self.assertEqual(geometry1, loss_dataframe['geometry'][0])
-        self.assertEqual('Colina', loss_dataframe['name'][0])
-        self.assertEqual(10, loss_dataframe['loss_value'][0])
-        self.assertEqual('$', loss_dataframe['loss_unit'][0])
-
-
-class DummyLossProvider():
-
-    #def get_loss(self, schema, taxonomy, from_damage_state, to_damage_state):
-    def get_loss(self, *args, **kwargs):
-        return 1
-
-    def get_unit(self):
-        return '$'
-
 
 
 if __name__ == '__main__':
